requirements.txt
absl-py==2.1.0
alembic==1.14.0
Cartopy==0.23.0
catboost==1.2.7
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.0.0
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
dask==2024.8.0
dask-expr==1.1.10
docker-pycreds==0.4.0
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.70.0
h5py==3.13.0
idna==3.10
imageio==2.36.1
importlib_resources==6.4.4
jax-cuda12-pjrt==0.4.30
jax-cuda12-plugin==0.4.30
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lightgbm==4.5.0
locket==1.0.0
Mako==1.3.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.8.93
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opt_einsum==3.4.0
optuna==4.1.0
pandas==2.2.2
partd==1.4.2
pillow==10.4.0
plotly==5.24.1
protobuf==5.28.2
pyarrow==17.0.0
PyKrige==1.7.2
pyparsing==3.1.4
pyproj==3.6.1
pyshp==2.3.1
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.16.0
setproctitle==1.3.3
shapely==2.0.6
smmap==5.0.1
SQLAlchemy==2.0.36
sympy==1.12
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2024.8.30
toolz==0.12.1
torch==2.4.0+cu121
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tqdm==4.66.5
triton==3.0.0
tzdata==2024.1
urllib3==2.2.3
wandb==0.18.3
Werkzeug==3.1.3
xgboost==2.1.1